fix deadlock issues in classificationstore, fieldcollections and relations table

+++ models/DataObject/Concrete/Dao.php
@@ -229,7 +229,11 @@ class Dao extends Model\DataObject\AbstractObject\Dao
             $condition = '(' . $condition . ' AND ownerType != "localizedfield" AND ownerType != "fieldcollection")';
         }
 
-        $this->db->deleteWhere('object_relations_' . $this->model->getClassId(), $condition);
+        $dataExists = $this->db->fetchOne("SELECT count(*) c FROM `object_relations_". $this->model->getClassId()."`
+        WHERE ".$condition);
+        if($dataExists) {
+            $this->db->deleteWhere('object_relations_' . $this->model->getClassId(), $condition);
+        }
 
         $inheritedValues = DataObject::doGetInheritedValues();
         DataObject::setGetInheritedValues(false);
